requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-import-export==4.3.7
djangorestframework==3.16.0
sqlparse==0.5.3
tablib==3.8.0
xlwt==1.3.0
